fix: SignUpContent 생성
