+++ app/src/main/java/org/sopt/at/presentation/ui/signup/SignUpActivity.kt
@@ -5,13 +5,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import org.sopt.at.presentation.ui.signin.SignInActivity
 import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
 
@@ -42,35 +35,4 @@ class SignUpActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun SignUpScreen(
-    modifier: Modifier = Modifier,
-    navigateToBack: () -> Unit = {},
-    navigateToSignIn: (String, String) -> Unit = { _, _ -> }
-) {
-    var currentStep by remember { mutableStateOf(1) }
-    var id by remember { mutableStateOf("") }
-    var pw by remember { mutableStateOf("") }
-
-    when (currentStep) {
-        1 -> IdInputScreen(
-            id = id,
-            onIdChange = { id = it },
-            onNext = { currentStep = 2 },
-            onBack = {
-                navigateToBack()
-            }
-        )
-
-        else -> PwInputScreen(
-            pw = pw,
-            onPwChange = { pw = it },
-            onNext = {
-                navigateToSignIn(id, pw)
-            }, onBack = { currentStep = 1 })
-    }
-
 }
